Improved compatibility with older compilers

+++ field_2d.h
@@ -461,8 +461,8 @@ class TabField: public TField{
 		void BField(long double x, long double y, long double z, long double t, long double B[4][4]){
 			long double r = sqrt(x*x+y*y);
 			long double Bscale = BFieldScale(t);
-			int indr = floor((r - r_mi)/rdist);
-			int indz = floor((z - z_mi)/zdist);
+			int indr = (int)floor((r - r_mi)/rdist);
+			int indz = (int)floor((z - z_mi)/zdist);
 			if (Bscale != 0 && indr >= 0 && indr < m - 1 && indz >= 0 && indz < n - 1){
 				// bicubic interpolation
 				long double rl = r_mi + indr*rdist;
@@ -513,8 +513,8 @@ class TabField: public TField{
 			long double dummy, Vloc, dVdrj[3];
 			if (Vc.nrows() > 0){ // prefer E-field from potential over pure E-field interpolation
 				long double r = sqrt(x*x+y*y);
-				int indr = floor((r - r_mi)/rdist);
-				int indz = floor((z - z_mi)/zdist);
+				int indr = (int)floor((r - r_mi)/rdist);
+				int indz = (int)floor((z - z_mi)/zdist);
 				if (indr >= 0 && indr < m - 1 && indz >= 0 && indz < n - 1){
 					// bicubic interpolation
 					long double rl = r_mi + indr*rdist;
@@ -529,8 +529,8 @@ class TabField: public TField{
 			}
 			else if (Erc.nrows() > 0 || Ephic.nrows() > 0 || Ezc.nrows() > 0){
 				long double r = sqrt(x*x+y*y);
-				int indr = floor((r - r_mi)/rdist);
-				int indz = floor((z - z_mi)/zdist);
+				int indr = (int)floor((r - r_mi)/rdist);
+				int indz = (int)floor((z - z_mi)/zdist);
 				if (indr >= 0 && indr < m - 1 && indz >= 0 && indz < n - 1){
 					// bicubic interpolation
 					long double rl = r_mi + indr*rdist;
